move require('alloy/styler').generateStyle() to Alloy.createStyle(), update test app
